backend/requirements.txt
Flask
Flask-Cors
pymongo
python-dotenv
flask-mail
gunicorn
tinydb
